type fix

+++ app/src/main/java/com/example/drivingstyleassistant/domain/entities/Events.java
@@ -92,7 +92,7 @@ public class Events {
         this.timestamp = timestamp;
     }
 
-    public int getRouteId() {
+    public long getRouteId() {
         return routeId;
     }
